impl(sidekick): move Config to separate package
Refactor the logic for creating and parsing the `sidekick.toml` config
file into a separate package. This change will prevent a dependency loop
in a future PR when introducing a `Config` struct to the `sample`
package for testing.
