+++ generator/internal/sidekick/command.go
@@ -19,11 +19,13 @@ import (
 	"fmt"
 	"slices"
 	"strings"
+
+	"github.com/googleapis/google-cloud-rust/generator/internal/config"
 )
 
 // command is an implementation of a sidekick command, like 'sidekick generate'.
 type command struct {
-	action           func(rootConfig *Config, cmdLine *CommandLine) error
+	action           func(rootConfig *config.Config, cmdLine *CommandLine) error
 	usageLine        string
 	altNames         []string
 	shortDescription string
@@ -115,7 +117,7 @@ func (c *command) visitAllFlags(fn func(f *flag.Flag)) {
 }
 
 // run executes the command's action, if it has one.
-func (c *command) run(rootConfig *Config, cmdLine *CommandLine) error {
+func (c *command) run(rootConfig *config.Config, cmdLine *CommandLine) error {
 	if c.action == nil {
 		return fmt.Errorf("command %s is not runnable", c.longName())
 	}
@@ -153,7 +155,7 @@ func newCommand(
 	shortDescription string,
 	longDescription string,
 	parent *command,
-	action func(rootConfig *Config, cmdLine *CommandLine) error,
+	action func(rootConfig *config.Config, cmdLine *CommandLine) error,
 ) *command {
 	if len(usageLine) == 0 {
 		panic("command usage line cannot be empty")
